Refactor the OAuthLogin extension to be part of the framework now. Not an extension anymore, but a module you can easily enable
