+++ public/api/v1/oauthlogin.php
@@ -1,10 +1,21 @@
 <?php
 
-use Ampersand\Extension\OAuthLogin\OAuthLoginController;
+use Ampersand\AmpersandApp;
+use Ampersand\SIAM\OAuth\LoginController;
 use Slim\Http\Request;
 use Slim\Http\Response;
-
 use function Ampersand\Misc\makeValidURL;
+
+// Adds menu item to login/logout
+/** @var \Ampersand\AngularApp $angularApp */
+global $angularApp;
+$angularApp->addMenuItem(
+    'role',
+    'app/src/oauthlogin/menuItem.html',
+    function (AmpersandApp $app) {
+        return $app->getSettings()->get('oauthlogin.enabled', false);
+    }
+);
 
 /**
  * @var \Slim\App $api
@@ -85,13 +96,13 @@ $api->group('/oauthlogin', function () {
         }
 
         // instantiate authController
-        $authController = new OAuthLoginController(
+        $authController = new LoginController(
             $identityProviders[$idp]['clientId'],
             $identityProviders[$idp]['clientSecret'],
             makeValidUrl($identityProviders[$idp]['redirectUrl'], $settings->get('global.serverURL')),
-            $identityProviders[$idp]['tokenUrl']
+            $identityProviders[$idp]['tokenUrl'],
+            $ampersandApp
         );
-        $authController->setAmpersandApp($ampersandApp);
 
         $api_url = $identityProviders[$idp]['apiUrl'];
         $isLoggedIn = $authController->authenticate($code, $idp, $api_url);
@@ -101,4 +112,17 @@ $api->group('/oauthlogin', function () {
 
         return $response->withRedirect($url);
     });
+})
+/**
+ * @phan-closure-scope \Slim\Container
+ */
+->add(function (Request $req, Response $res, callable $next) {
+    /** @var \Ampersand\AmpersandApp $ampersandApp */
+    $ampersandApp = $this['ampersand_app'];
+
+    if (!$ampersandApp->getSettings()->get('oauthlogin.enabled', false)) {
+        throw new Exception("The OAuth login module is not enabled", 501);
+    }
+
+    return $next($req, $res);
 });
