+++ public/app/images/oauth/logo-github.png
